# backend/internal/api/docs/configer_swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Write-back-native GitOps configuration management. Configer discovers an existing repository's layout, exposes a parameter x instance grid over the repository's OWN files, and surgically edits them; every change flows to Git as commits, branches, pull requests, and merges. `.configer/` holds metadata only.\n\n## Repository scoping\nEvery resource endpoint exists in two shapes: the unscoped `/api/...` form acts on the default (first-connected) repository for single-repo deployments and older clients, and the identical routes are available per repository under `/api/repos/{repoId}/...`. Only the unscoped forms are documented here; to target a specific repository prefix any path with `/api/repos/{repoId}`.\n\n## Errors\nEvery 4xx/5xx response uses one envelope: a stable machine-readable `code`, a human-readable `error`, a `requestId` (also returned as the `X-Request-ID` header) that correlates to server logs, and - for validation failures - a `fields` array. Branch on `code`, never on the message text.\n\n## Correlation\nSend `X-Request-ID` to have it echoed and used as the correlation id; otherwise the server generates one. It is returned on every response.",
        "title": "Configer API",
...
        "contact": {
            "name": "Configer",
...
        },
...
    },
    "basePath": "/",
    "paths": {
...
        "/api/changes": {
            "get": {
                "description": "Change requests in all states (Draft, UnderReview, Approved, Published, Rejected), newest first. Cursor-paginated: pass `limit` (default 50, max 200) and the previous response's `nextCursor`. Returns `{items, nextCursor, hasMore}`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "List change requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque cursor from the previous page",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Page-change_ChangeRequest"
                        }
                    }
                }
            }
        },
...
        "/api/changes/{id}": {
            "get": {
                "description": "One change request by id. Refreshes the linked pull-request state (merged/closed/checks) from the host before returning.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Get a change request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Unknown change request",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/approve": {
            "post": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Record approver sign-off on an under-review change request, advancing it to Approved without publishing (Publish/merge is a separate step). Requires the approver role when auth is enabled. 409 if not under review.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Approve a change request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "403": {
                        "description": "Approver role required",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "409": {
                        "description": "Not under review",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/comments": {
            "post": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Append a review note. The session identity is the author (the body's author field is only a single-user-mode fallback).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Comment on a change request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The comment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id, body, or empty comment",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Unknown change request",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/merge": {
            "post": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Approve and merge the change request's pull request, publishing it. Requires the approver role when auth is enabled. Async: returns 202 with the change resource; poll `state` for Published. 409 if not mergeable; 502/504 on a downstream failure.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Merge (publish) a change request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "403": {
                        "description": "Approver role required",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "409": {
                        "description": "Not in a mergeable state",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "502": {
                        "description": "The merge failed upstream",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "504": {
                        "description": "The merge timed out upstream",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/pr-status": {
            "get": {
                "description": "Live CI check roll-up (passing/failing/pending/none) and host merge-readiness for a change request's pull request, read fresh from the provider. Returns `{supported:false}` when the change has no hosted PR (pure-git deployment or not yet submitted).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Pull-request status (CI + mergeability)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Unknown change request",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/preview": {
            "get": {
                "description": "The byte-level plan for a change request: for every file its value and file edits would rewrite, the exact content before and after (so the UI can show a real diff), plus +/- line counts and one-line summaries of any structural instance changes. Builds the edits in a throwaway checkout of the base branch; nothing is committed or pushed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Preview a change request's file edits",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Unknown change request",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "502": {
                        "description": "A downstream (git) step failed",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/changes/{id}/reject": {
            "post": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Reject/close the change request (a draft is discarded, a submitted one is closed). An optional reason is recorded as a review comment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Reject a change request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Change request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "409": {
                        "description": "Not in a rejectable state",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "502": {
                        "description": "Closing the pull request failed upstream",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CommentRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "example": "Looks good, ship after the window."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieSession": {
            "description": "Session cookie set by the GitHub OAuth login flow (`GET /api/auth/login` -\u003e callback). Only required when the deployment has OAuth configured; single-user deployments need no authentication.",
            "type": "apiKey",
            "name": "configer_session",
            "in": "cookie"
        }
    }
}